core/local/chokidar: Extract finalSort() from analysis
